refs #18 controllerとメソッドを動的に呼び出す仕組みを構築
