+++ controllers/root_controller.go
@@ -1,8 +1,22 @@
 package controllers
 import (
+	"fmt"
 	"net/http"
 	"github.com/zenazn/goji/web"
+	"github.com/flosch/pongo2"
 )
 
-func Index(c web.C, w http.ResponseWriter, r *http.Request) {
+type Root struct {
+}
+func (u *Root)Index(c web.C, w http.ResponseWriter, r *http.Request) {
+	current_user, result := LoginRequired(c, w, r)
+	if result {
+		fmt.Printf("current_user: %+v\n", *current_user)
+	}
+	tpl, err := pongo2.DefaultSet.FromFile("views/home.html.tpl")
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	tpl.ExecuteWriter(pongo2.Context{"title": "Fascia"}, w)
 }
